Added spawn base system/prefab.  Split Spawner class into three.  Added Activate Spawner Endeavour.  Made Health scripts more flexible.
