+++ Assets/Prefabs/GameControl/SpawnBase.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 369faafd6aa2d00408f431797e35478a
+timeCreated: 1464302787
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
