+++ Assets/Scripts/Attributes/Endeavour/Factories/ActivateSpawner.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using System.Collections;
+
+[System.Serializable]
+public class ActivateSpawner : EndeavourFactory {
+	
+	public override Endeavour constructEndeavour(RobotController controller) {
+		if (parent == null) {
+			return null;
+		}
+		return new ActivateSpawnerAction(controller, goals, parent.labelHandle, parent.GetComponent<RobotSpawner>());
+	}
+}
