+++ Assets/Scripts/Robots/AbstractRobotSpawner.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 94928ab5cbc9b104d851892a66cf9d03
-timeCreated: 1464286031
+guid: a66636efdf3f51a4c8a771aa2c0b058a
+timeCreated: 1464286369
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
